+++ Web/Twitter/Conduit/Base.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE RecordWildCards #-}
 
 module Web.Twitter.Conduit.Base
@@ -16,7 +15,6 @@ module Web.Twitter.Conduit.Base
        , sourceWithCursor'
        , sourceWithSearchResult
        , sourceWithSearchResult'
-       , TwitterBaseM
        , endpoint
        , makeRequest
        , sinkJSON
@@ -24,16 +22,16 @@ module Web.Twitter.Conduit.Base
        ) where
 
 import Web.Twitter.Conduit.Cursor
-import Web.Twitter.Conduit.Parameters
+import Web.Twitter.Conduit.Parameters hiding (url)
 import Web.Twitter.Conduit.Request
 import Web.Twitter.Conduit.Response
 import Web.Twitter.Conduit.Types
 import Web.Twitter.Types.Lens
 
 import Control.Lens
-import Control.Monad.IO.Class
-import Control.Monad.Trans.Class (lift)
-import Control.Monad.Trans.Resource (MonadResource, MonadThrow, monadThrow)
+import Control.Monad.Base
+import Control.Monad.Catch (MonadThrow (..))
+import Control.Monad.Trans.Resource (MonadResource, ResourceT, runResourceT)
 import Data.Aeson
 import Data.Aeson.Lens
 import Data.ByteString (ByteString)
@@ -48,9 +46,6 @@ import qualified Network.HTTP.Conduit as HTTP
 import qualified Network.HTTP.Types as HT
 import Unsafe.Coerce
 import Web.Authenticate.OAuth (signOAuth)
-
-type TwitterBaseM m = ( MonadResource m
-                      )
 
 makeRequest :: APIRequest apiName responseType
             -> IO HTTP.Request
@@ -68,10 +63,13 @@ makeRequest' :: HT.Method -- ^ HTTP request method (GET or POST)
              -> IO HTTP.Request
 makeRequest' m url query = do
     req <- HTTP.parseUrl url
-    return $ req { HTTP.method = m
-                 , HTTP.queryString = HT.renderSimpleQuery False query
-                 , HTTP.checkStatus = \_ _ _ -> Nothing
-                 }
+    let addParams =
+            if m == "POST"
+            then HTTP.urlEncodedBody query
+            else \r -> r { HTTP.queryString = HT.renderSimpleQuery False query }
+    return $ addParams $ req { HTTP.method = m
+                             , HTTP.checkStatus = \_ _ _ -> Nothing
+                             }
 
 getResponse :: MonadResource m
             => TWInfo
@@ -90,9 +88,8 @@ getResponse TWInfo{..} mgr req = do
 endpoint :: String
 endpoint = "https://api.twitter.com/1.1/"
 
-getValue :: MonadThrow m
-         => Response (C.ResumableSource m ByteString)
-         -> m (Response Value)
+getValue :: Response (C.ResumableSource (ResourceT IO) ByteString)
+         -> ResourceT IO (Response Value)
 getValue res = do
     value <- responseBody res C.$$+- sinkJSON
     return $ res { responseBody = value }
@@ -114,17 +111,17 @@ checkResponse Response{..} =
   where
     sci = HT.statusCode responseStatus
 
-getValueOrThrow :: (MonadThrow m, FromJSON a)
-                => Response (C.ResumableSource m ByteString)
-                -> m (Response a)
+getValueOrThrow :: FromJSON a
+                => Response (C.ResumableSource (ResourceT IO) ByteString)
+                -> ResourceT IO (Response a)
 getValueOrThrow res = do
     res' <- getValue res
     case checkResponse res' of
-        Left err -> monadThrow err
+        Left err -> throwM err
         Right _ -> return ()
     case fromJSON (responseBody res') of
         Success r -> return $ res' { responseBody = r }
-        Error err -> monadThrow $ FromJSONError err
+        Error err -> throwM $ FromJSONError err
 
 -- | Perform an 'APIRequest' and then provide the response which is mapped to a suitable type of
 -- <http://hackage.haskell.org/package/twitter-types twitter-types>.
@@ -132,28 +129,27 @@ getValueOrThrow res = do
 -- Example:
 --
 -- @
--- 'HTTP.withManager' $ \\mgr -\> do
---      user <- 'call' twInfo mgr $ 'accountVerifyCredentials'
---      'liftIO' $ print user
+-- user <- 'call' twInfo mgr $ 'accountVerifyCredentials'
+-- print user
 -- @
 --
 -- If you need raw JSON value which is parsed by <http://hackage.haskell.org/package/aeson aeson>,
 -- use 'call'' to obtain it.
-call :: (MonadResource m, FromJSON responseType)
+call :: FromJSON responseType
      => TWInfo -- ^ Twitter Setting
      -> HTTP.Manager
      -> APIRequest apiName responseType
-     -> m responseType
+     -> IO responseType
 call = call'
 
 -- | Perform an 'APIRequest' and then provide the response.
 -- The response of this function is not restrict to @responseType@,
 -- so you can choose an arbitrarily type of FromJSON instances.
-call' :: (MonadResource m, FromJSON value)
+call' :: FromJSON value
       => TWInfo -- ^ Twitter Setting
       -> HTTP.Manager
       -> APIRequest apiName responseType
-      -> m value
+      -> IO value
 call' info mgr req = responseBody `fmap` callWithResponse' info mgr req
 
 -- | Perform an 'APIRequest' and then provide the 'Response'.
@@ -161,17 +157,16 @@ call' info mgr req = responseBody `fmap` callWithResponse' info mgr req
 -- Example:
 --
 -- @
--- res \<- 'HTTP.withManager' $ \\mgr -\> do
---     'callWithResponse' twInfo mgr $ 'accountVerifyCredentials'
+-- res \<- 'callWithResponse' twInfo mgr $ 'accountVerifyCredentials'
 -- 'print' $ 'responseStatus' res
 -- 'print' $ 'responseHeaders' res
 -- 'print' $ 'responseBody' res
 -- @
-callWithResponse :: (MonadResource m, FromJSON responseType)
+callWithResponse :: FromJSON responseType
                  => TWInfo -- ^ Twitter Setting
                  -> HTTP.Manager
                  -> APIRequest apiName responseType
-                 -> m (Response responseType)
+                 -> IO (Response responseType)
 callWithResponse = callWithResponse'
 
 -- | Perform an 'APIRequest' and then provide the 'Response'.
@@ -181,25 +176,25 @@ callWithResponse = callWithResponse'
 -- Example:
 --
 -- @
--- res \<- 'HTTP.withManager' $ \\mgr -\> do
---     'callWithResponse'' twInfo mgr $ 'accountVerifyCredentials'
+-- res \<- 'callWithResponse'' twInfo mgr $ 'accountVerifyCredentials'
 -- 'print' $ 'responseStatus' res
 -- 'print' $ 'responseHeaders' res
 -- 'print' $ 'responseBody' (res :: Value)
 -- @
-callWithResponse' :: (MonadResource m, FromJSON value)
+callWithResponse' :: FromJSON value
                   => TWInfo
                   -> HTTP.Manager
                   -> APIRequest apiName responseType
-                  -> m (Response value)
-callWithResponse' info mgr req = do
-    res <- getResponse info mgr =<< liftIO (makeRequest req)
-    getValueOrThrow res
+                  -> IO (Response value)
+callWithResponse' info mgr req =
+    runResourceT $ do
+        res <- getResponse info mgr =<< liftBase (makeRequest req)
+        getValueOrThrow res
 
 -- | A wrapper function to perform multiple API request with changing @max_id@ parameter.
 --
 -- This function cooperate with instances of 'HasMaxIdParam'.
-sourceWithMaxId :: ( MonadResource m
+sourceWithMaxId :: ( MonadBase IO m
                    , FromJSON responseType
                    , AsStatus responseType
                    , HasMaxIdParam (APIRequest apiName [responseType])
@@ -211,7 +206,7 @@ sourceWithMaxId :: ( MonadResource m
 sourceWithMaxId info mgr = loop
   where
     loop req = do
-        res <- lift $ call info mgr req
+        res <- liftBase $ call info mgr req
         case getMinId res of
             Just mid -> do
                 CL.sourceList res
@@ -224,8 +219,8 @@ sourceWithMaxId info mgr = loop
 -- so you can choose an arbitrarily type of FromJSON instances.
 --
 -- This function cooperate with instances of 'HasMaxIdParam'.
-sourceWithMaxId' :: ( MonadResource m
-                    , HasMaxIdParam (APIRequest apiName [responseType])
+sourceWithMaxId' :: ( MonadBase IO m
+                    ,  HasMaxIdParam (APIRequest apiName [responseType])
                     )
                  => TWInfo -- ^ Twitter Setting
                  -> HTTP.Manager
@@ -234,7 +229,7 @@ sourceWithMaxId' :: ( MonadResource m
 sourceWithMaxId' info mgr = loop
   where
     loop req = do
-        res <- lift $ call' info mgr req
+        res <- liftBase $ call' info mgr req
         case getMinId res of
             Just mid -> do
                 CL.sourceList res
@@ -245,7 +240,7 @@ sourceWithMaxId' info mgr = loop
 -- | A wrapper function to perform multiple API request with changing @cursor@ parameter.
 --
 -- This function cooperate with instances of 'HasCursorParam'.
-sourceWithCursor :: ( MonadResource m
+sourceWithCursor :: ( MonadBase IO m
                     , FromJSON responseType
                     , CursorKey ck
                     , HasCursorParam (APIRequest apiName (WithCursor ck responseType))
@@ -258,7 +253,7 @@ sourceWithCursor info mgr req = loop (-1)
   where
     loop 0 = CL.sourceNull
     loop cur = do
-        res <- lift $ call info mgr $ req & cursor ?~ cur
+        res <- liftBase $ call info mgr $ req & cursor ?~ cur
         CL.sourceList $ contents res
         loop $ nextCursor res
 
@@ -267,7 +262,7 @@ sourceWithCursor info mgr req = loop (-1)
 -- so you can choose an arbitrarily type of FromJSON instances.
 --
 -- This function cooperate with instances of 'HasCursorParam'.
-sourceWithCursor' :: ( MonadResource m
+sourceWithCursor' :: ( MonadBase IO m
                      , FromJSON responseType
                      , CursorKey ck
                      , HasCursorParam (APIRequest apiName (WithCursor ck responseType))
@@ -284,12 +279,12 @@ sourceWithCursor' info mgr req = loop (-1)
     relax = unsafeCoerce
     loop 0 = CL.sourceNull
     loop cur = do
-        res <- lift $ call info mgr $ relax $ req & cursor ?~ cur
+        res <- liftBase $ call info mgr $ relax $ req & cursor ?~ cur
         CL.sourceList $ contents res
         loop $ nextCursor res
 
 -- | A wrapper function to perform multiple API request with @SearchResult@.
-sourceWithSearchResult :: ( MonadResource m
+sourceWithSearchResult :: ( MonadBase IO m
                           , FromJSON responseType
                           , HasMaxIdParam (APIRequest apiName (SearchResult [responseType]))
                           )
@@ -298,7 +293,7 @@ sourceWithSearchResult :: ( MonadResource m
                        -> APIRequest apiName (SearchResult [responseType])
                        -> m (SearchResult (C.Source m responseType))
 sourceWithSearchResult info mgr req = do
-    res <- call info mgr req
+    res <- liftBase $ call info mgr req
     let body = CL.sourceList (res ^. searchResultStatuses) <>
                loop (res ^. searchResultSearchMetadata . searchMetadataNextResults)
     return $ res & searchResultStatuses .~ body
@@ -308,12 +303,12 @@ sourceWithSearchResult info mgr req = do
     loop (Just nextResultsStr) = do
         let nextResults = nextResultsStr & HT.parseSimpleQuery . T.encodeUtf8 & traversed . _2 %~ (PVString . T.decodeUtf8)
             nextParams = M.toList $ M.union (M.fromList nextResults) origQueryMap
-        res <- call info mgr $ req & params .~ nextParams
+        res <- liftBase $ call info mgr $ req & params .~ nextParams
         CL.sourceList (res ^. searchResultStatuses)
         loop $ res ^. searchResultSearchMetadata . searchMetadataNextResults
 
 -- | A wrapper function to perform multiple API request with @SearchResult@.
-sourceWithSearchResult' :: ( MonadResource m
+sourceWithSearchResult' :: ( MonadBase IO m
                            , HasMaxIdParam (APIRequest apiName (SearchResult [responseType]))
                            )
                         => TWInfo -- ^ Twitter Setting
@@ -321,7 +316,7 @@ sourceWithSearchResult' :: ( MonadResource m
                         -> APIRequest apiName (SearchResult [responseType])
                         -> m (SearchResult (C.Source m Value))
 sourceWithSearchResult' info mgr req = do
-    res <- call info mgr $ relax req
+    res <- liftBase $ call info mgr $ relax req
     let body = CL.sourceList (res ^. searchResultStatuses) <>
                loop (res ^. searchResultSearchMetadata . searchMetadataNextResults)
     return $ res & searchResultStatuses .~ body
@@ -335,7 +330,7 @@ sourceWithSearchResult' info mgr req = do
     loop (Just nextResultsStr) = do
         let nextResults = nextResultsStr & HT.parseSimpleQuery . T.encodeUtf8 & traversed . _2 %~ (PVString . T.decodeUtf8)
             nextParams = M.toList $ M.union (M.fromList nextResults) origQueryMap
-        res <- call info mgr $ relax $ req & params .~ nextParams
+        res <- liftBase $ call info mgr $ relax $ req & params .~ nextParams
         CL.sourceList (res ^. searchResultStatuses)
         loop $ res ^. searchResultSearchMetadata . searchMetadataNextResults
 
@@ -349,5 +344,5 @@ sinkFromJSON :: ( FromJSON a
 sinkFromJSON = do
     v <- sinkJSON
     case fromJSON v of
-        Error err -> monadThrow $ FromJSONError err
+        Error err -> throwM $ FromJSONError err
         Success r -> return r